fix(digest): parse.pyを責務分割し200行制約とコメント密度規約を満たす
レビューで指摘されたImportant 2件（1ファイル200行超過・コメント密度不足）を
YAML読み取り/ブロック分割/finding変換/データモデル/エラー型の責務単位で分割して解消。
digest_md.parseから5シンボル(DigestError/Document/Finding/parse_document/
parse_yaml_block)をそのままimportできる契約は維持。あわせてMinor指摘の
分岐網羅テスト9件（suppressed経路・不正category/severity・重複slug・
yamlブロック欠落・未閉フェンス・未知見出し）を追加。
